+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All Notable changes to `gmponos/guzzle_logger` will be documented in this file
 
+## 1.0.0 - NEXT RELEASE
+
+### Changes
+- Added more classes of status codes to `ThresholdStrategy`
+
 ## 0.8.0 - 2018-12-13
 
 ### Changes